Refactor component packages and add theme support for UI elements
