+++ components/src/desktopMain/kotlin/br/ufpa/genewhale/buttons/MyOutlinedButton.kt
@@ -1,0 +1,26 @@
+package br.ufpa.genewhale.buttons
+
+import androidx.compose.foundation.layout.RowScope
+import androidx.compose.material3.OutlinedButton
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.input.pointer.PointerIcon
+import androidx.compose.ui.input.pointer.pointerHoverIcon
+import br.ufpa.genewhale.theme.ThemeDefaults
+
+@Composable
+fun MyOutlinedButton(
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    onClick: () -> Unit,
+    content: @Composable RowScope.() -> Unit
+) {
+    val otherModifier = if (enabled) Modifier.pointerHoverIcon(PointerIcon.Hand) else Modifier
+    OutlinedButton(
+        modifier = modifier.then(otherModifier),
+        onClick = onClick,
+        enabled = enabled,
+        shape = ThemeDefaults.ButtonShape,
+        content = content
+    )
+}
