+++ components/src/desktopMain/kotlin/br/ufpa/genewhale/buttons/MyTab.kt
@@ -1,0 +1,112 @@
+package br.ufpa.genewhale.buttons
+
+import androidx.compose.animation.animateColor
+import androidx.compose.animation.core.LinearEasing
+import androidx.compose.animation.core.tween
+import androidx.compose.animation.core.updateTransition
+import androidx.compose.foundation.Indication
+import androidx.compose.foundation.interaction.MutableInteractionSource
+import androidx.compose.foundation.layout.Arrangement
+import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.ColumnScope
+import androidx.compose.foundation.layout.fillMaxWidth
+import androidx.compose.foundation.selection.selectable
+import androidx.compose.material.ripple.rememberRipple
+import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.LocalContentColor
+import androidx.compose.material3.LocalUseFallbackRippleImplementation
+import androidx.compose.material3.ripple
+import androidx.compose.runtime.Composable
+import androidx.compose.runtime.CompositionLocalProvider
+import androidx.compose.runtime.getValue
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.input.pointer.PointerIcon
+import androidx.compose.ui.input.pointer.pointerHoverIcon
+import androidx.compose.ui.semantics.Role
+import androidx.compose.ui.unit.Dp
+
+@Composable
+fun MyTab(
+    selected: Boolean,
+    onClick: () -> Unit,
+    modifier: Modifier = Modifier,
+    enabled: Boolean = true,
+    selectedContentColor: Color = LocalContentColor.current,
+    unselectedContentColor: Color = selectedContentColor,
+    interactionSource: MutableInteractionSource? = null,
+    content: @Composable ColumnScope.() -> Unit
+) {
+    val otherModifier = if (enabled) Modifier.pointerHoverIcon(PointerIcon.Hand) else Modifier
+    // The color of the Ripple should always the selected color, as we want to show the color
+    // before the item is considered selected, and hence before the new contentColor is
+    // provided by TabTransition.
+    val ripple = rippleOrFallbackImplementation(bounded = true, color = selectedContentColor)
+
+    TabTransition(selectedContentColor, unselectedContentColor, selected) {
+        Column(
+            modifier =
+                modifier
+                    .then(otherModifier)
+                    .selectable(
+                        selected = selected,
+                        onClick = onClick,
+                        enabled = enabled,
+                        role = Role.Tab,
+                        interactionSource = interactionSource,
+                        indication = ripple
+                    )
+                    .fillMaxWidth(),
+            horizontalAlignment = Alignment.CenterHorizontally,
+            verticalArrangement = Arrangement.Center,
+            content = content
+        )
+    }
+}
+
+@Composable
+private fun TabTransition(
+    activeColor: Color,
+    inactiveColor: Color,
+    selected: Boolean,
+    content: @Composable () -> Unit
+) {
+    val transition = updateTransition(selected)
+    val color by
+    transition.animateColor(
+        transitionSpec = {
+            if (false isTransitioningTo true) {
+                tween(
+                    durationMillis = TabFadeInAnimationDuration,
+                    delayMillis = TabFadeInAnimationDelay,
+                    easing = LinearEasing
+                )
+            } else {
+                tween(durationMillis = TabFadeOutAnimationDuration, easing = LinearEasing)
+            }
+        }
+    ) {
+        if (it) activeColor else inactiveColor
+    }
+    CompositionLocalProvider(LocalContentColor provides color, content = content)
+}
+
+private const val TabFadeInAnimationDuration = 150
+private const val TabFadeInAnimationDelay = 100
+private const val TabFadeOutAnimationDuration = 100
+
+@Suppress("DEPRECATION_ERROR")
+@OptIn(ExperimentalMaterial3Api::class)
+@Composable
+internal fun rippleOrFallbackImplementation(
+    bounded: Boolean = true,
+    radius: Dp = Dp.Unspecified,
+    color: Color = Color.Unspecified
+): Indication {
+    return if (LocalUseFallbackRippleImplementation.current) {
+        rememberRipple(bounded, radius, color)
+    } else {
+        ripple(bounded, radius, color)
+    }
+}
