+++ components/src/desktopMain/kotlin/br/ufpa/genewhale/theme/GenomeTheme.kt
@@ -1,0 +1,16 @@
+package br.ufpa.genewhale.theme
+
+import androidx.compose.material3.ColorScheme
+import androidx.compose.material3.MaterialTheme
+import androidx.compose.runtime.Composable
+
+@Composable
+fun GenomeTheme(
+    colorScheme: ColorScheme = MyColorScheme,
+    content: @Composable () -> Unit
+) {
+    MaterialTheme(
+        content = content,
+        colorScheme = colorScheme
+    )
+}
