--- composeApp/src/desktopMain/kotlin/br/ufpa/genewhale/ui/components/dropdown/DropdownMenu.kt
@@ -1,2 +1,0 @@
-package br.ufpa.genewhale.ui.components.dropdown
-
