+++ composeApp/src/desktopMain/kotlin/br/ufpa/genewhale/ui/screens/Home.kt
@@ -13,8 +13,8 @@ import androidx.compose.runtime.Composable
 import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.unit.dp
-import br.ufpa.genewhale.ui.components.cards.ToolCard
-import br.ufpa.genewhale.ui.components.cards.Tools
+import br.ufpa.genewhale.cards.ToolCard
+import br.ufpa.genewhale.cards.Tools
 import br.ufpa.genewhale.ui.states.HomeUIState
 import br.ufpa.genewhale.ui.states.HomeUiIntent
 import br.ufpa.genewhale.ui.theme.GenomeTheme
