+++ composeApp/src/desktopMain/kotlin/br/ufpa/genewhale/ui/screens/tools/Panaroo.kt
@@ -17,18 +17,18 @@ import androidx.compose.ui.draw.clip
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.dp
 import androidx.compose.ui.unit.sp
-import br.ufpa.genewhale.ui.components.Terminal
-import br.ufpa.genewhale.ui.components.buttons.MyButton
-import br.ufpa.genewhale.ui.components.buttons.MyOutlinedButton
-import br.ufpa.genewhale.ui.components.buttons.MyTab
-import br.ufpa.genewhale.ui.components.buttons.RoundedCornerCheckBox
-import br.ufpa.genewhale.ui.components.clickableWithHoverIcon
-import br.ufpa.genewhale.ui.components.dropdown.DropdownSelector
-import br.ufpa.genewhale.ui.components.pickfiles.PickFolder
-import br.ufpa.genewhale.ui.components.sliders.MemorySlider
-import br.ufpa.genewhale.ui.components.sliders.ThreadsSlider
-import br.ufpa.genewhale.ui.components.textfield.CustomTextField
-import br.ufpa.genewhale.ui.components.tooltips.MyTooltip
+import br.ufpa.genewhale.Terminal
+import br.ufpa.genewhale.buttons.MyButton
+import br.ufpa.genewhale.buttons.MyOutlinedButton
+import br.ufpa.genewhale.buttons.MyTab
+import br.ufpa.genewhale.buttons.RoundedCornerCheckBox
+import br.ufpa.genewhale.clickableWithHoverIcon
+import br.ufpa.genewhale.dropdown.DropdownSelector
+import br.ufpa.genewhale.pickfiles.PickFolder
+import br.ufpa.genewhale.sliders.MemorySlider
+import br.ufpa.genewhale.sliders.ThreadsSlider
+import br.ufpa.genewhale.textfield.CustomTextField
+import br.ufpa.genewhale.tooltips.MyTooltip
 import br.ufpa.genewhale.ui.states.tools.*
 import br.ufpa.genewhale.ui.theme.GenomeTheme
 import io.github.vinceglb.filekit.path
